core/src/test/java/com/avolution/net/udp/UDPNettyServiceTest.java
package com.avolution.net.udp;

import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;

import static org.junit.jupiter.api.Assertions.*;

class UDPNettyServiceTest {

    private UDPNettyService udpNettyService;

    @BeforeEach
    void setUp() {
        udpNettyService = new UDPNettyService(8080);
    }

    @Test
    void testStartService() {
        udpNettyService.start();
        assertTrue(udpNettyService.isRunning());
    }

    @Test
    void testStopService() {
        udpNettyService.start();
        udpNettyService.stop();
        assertFalse(udpNettyService.isRunning());
    }

    @Test
    void testRestartService() {
        udpNettyService.start();
        udpNettyService.restart();
        assertTrue(udpNettyService.isRunning());
    }

    @Test
    void testHandleIncomingPacket() {
        // Simulate incoming packet and verify handling
        byte[] content = "Test packet".getBytes();
        UDPPacket packet = new UDPPacket(1, 0, content);
        // Add logic to simulate incoming packet and verify handling
    }

    @Test
    void testHandleOutgoingPacket() {
        // Simulate outgoing packet and verify handling
        byte[] content = "Test packet".getBytes();
        UDPPacket packet = new UDPPacket(1, 0, content);
        // Add logic to simulate outgoing packet and verify handling
    }
}
